Group5_Main/backend_GrpNo.5/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.1.1
Flask-JWT-Extended>=4.6.0
ultralytics>=8.1.0
opencv-python>=4.8.1.78
numpy>=1.26.0
Pillow>=10.1.0
python-dotenv>=1.0.0
Werkzeug>=3.0.1
flask-socketio>=5.3.6
python-socketio>=5.10.0
bcrypt>=4.1.2
setuptools>=69.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9